separate state migration code from blockchain.go
